% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{carga_paquetes}
\alias{carga_paquetes}
\title{carga_paquetes}
\usage{
carga_paquetes()
}
\description{
Esta función carga los paquetes comúnmente utilizados para análisis de datos, visualización y modelado estadístico.
}
\examples{
carga_paquetes()

}
