% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connection}
\alias{connection}
\title{Ejecutar una consulta SQL en una base de datos}
\usage{
connection(query, db)
}
\arguments{
\item{query}{Cadena de texto con la consulta SQL que se desea ejecutar.}

\item{db}{Número entero: \code{1} para la base de datos "Banco", \code{2} para "Seguros".}
}
\value{
Un dataframe con los resultados de la consulta o \code{NULL} si ocurre un error.
}
\description{
Se conecta a una base de datos SQL Server mediante ODBC y ejecuta una consulta SQL.
}
\examples{
\dontrun{
result <- connection("SELECT * FROM tabla", db = 1)
}
}
