% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\docType{data}
\name{credentials}
\alias{credentials}
\title{Lista de credenciales para bases de datos}
\format{
An object of class \code{list} of length 1.
}
\usage{
credentials
}
\description{
Contiene la información de conexión a las bases de datos "Banco" y "Seguros".
Se usa internamente en la función \code{connection()}.
}
\keyword{internal}
