% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{elbow_test}
\alias{elbow_test}
\title{Prueba del Codo}
\usage{
elbow_test(data, cluster_range = 1:10, output_name = NULL)
}
\arguments{
\item{data}{Dataframe para el clustering}

\item{cluster_range}{n vector numérico con los valores de \code{k} a evaluar (por defecto 1:10)}

\item{output_name}{Nombre del archivo PDF donde guardar el gráfico (opcional)}
}
\value{
Un gráfico mostrando la inercia vs número de clusters
}
\description{
Hacer la gráfica para la prueba del codo para un dataset
}
\examples{
\dontrun{
elbow_test(data, cluster_range = 1:10)
}
}
