% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_lots_excels}
\alias{read_lots_excels}
\title{Lectura de Muchos Excels}
\usage{
read_lots_excels(path)
}
\arguments{
\item{path}{Ruta a la carpeta que se desea utilizar.}
}
\value{
lista de todos los dataframes leídos.
}
\description{
Función que lee todos los archivos de excel en una ruta dada.
}
\examples{
\dontrun{
dfs_list <- read_lots_excels('path/to/directory')
}
}
