% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{redondear}
\alias{redondear}
\title{Función de redondeo}
\usage{
redondear(monto)
}
\arguments{
\item{monto}{Número a redondear}
}
\value{
Un entero redondeado
}
\description{
Función que toma un número entero o decimal y lo redondea según ciertas reglas y según su orden de magnitud.
}
\examples{
\dontrun{
redondeado <- redondear(monto)
}
}
