% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{scatter_cluster}
\alias{scatter_cluster}
\title{Scatterplot de Clusters}
\usage{
scatter_cluster(df, x_col, y_col, cluster_col = "Cluster", output_name = NULL)
}
\arguments{
\item{df}{Dataframe segmentado en clusters}

\item{x_col}{nombre de la columna a colocar en el eje x}

\item{y_col}{nombre de la columna a colocar en el eje y}

\item{cluster_col}{Nombre de la columna en la que se encuentra la segmentacion por cluster (por defecto 'Cluster')}

\item{output_name}{Nombre del archivo PDF donde guardar el gráfico (opcional)}
}
\value{
Un gráfico mostrando las columnas ingresadas.
}
\description{
Funcion que grafica en 2D los clusteres generados con kmeans
}
\examples{
\dontrun{
scatter_cluster(df, "eje x", "eje y", "Cluster")
}
}
