% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{table_percentiles}
\alias{table_percentiles}
\title{Tabla con los Percentiles de un Dataset}
\usage{
table_percentiles(df, columna, pasos = 1)
}
\arguments{
\item{df}{Dataframe donde se encuentra la columna a utilizar.}

\item{columna}{Nombre de la columna a analizar.}

\item{pasos}{Rango de la división de percentiles (por defecto 1).}
}
\value{
Un dataframe con los percentiles de la columna proporcionada
}
\description{
Genera una tabla con los percentiles de un dataframe proporcionado
}
\examples{
\dontrun{
result <- table_percentiles(df, "valor")
}
}
