% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tukey}
\alias{tukey}
\title{Método de Tukey}
\usage{
tukey(df, varColumn, outputColumn = "outlier", c = 1.5)
}
\arguments{
\item{df}{Dataframe}

\item{varColumn}{Nombre de la columna a analizar.}

\item{c}{Factor de ajuste (por defecto 1.5).}

\item{outputColumne}{Nombre de la columna en la que se categorizan como outliers o no (por defecto 'outlier').}
}
\value{
Dataframe original con una nueva columna que indica si un valor es un outlier (1) o no (0).
}
\description{
Función que implementa el método de tukey por rango intercuartílico para la detección de outliers.
}
\examples{
\dontrun{
resultado <- tukey(df, "valor")
}
}
