% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tukey_alternative}
\alias{tukey_alternative}
\title{Método Alternativo de Tukey}
\usage{
tukey_alternative(
  df,
  varColumn,
  outputColumn = "outlier",
  c = 1.5,
  low_percentile = 0.05,
  high_percentile = 0.95,
  filter = TRUE
)
}
\arguments{
\item{df}{Dataframe}

\item{varColumn}{Nombre de la columna a analizar.}

\item{c}{Factor de ajuste (por defecto 1.5).}

\item{low_percentile}{Percentil inferior para la filtración de datos antes del análisis (por defecto 0.05).}

\item{high_percentile}{Percentil superior para la filtración de datos antes del análisis (por defecto 0.95).}

\item{filter}{Booleano. Si es TRUE, se filtran los datos dentro del rango de percentiles antes de calcular los límites (por defecto TRUE).}

\item{outputColumne}{Nombre de la columna en la que se categorizan como outliers o no (por defecto 'outlier').}
}
\value{
Dataframe original con una nueva columna que indica si un valor es un outlier (1) o no (0).
}
\description{
Otra implementación del método de tukey, basada en la media superior e inferior.
}
\examples{
\dontrun{
resultado <- tukey_alternative(df, "valor")
}
}
